Fix printf format warnings when compiling on Windows.

+++ src/seg006.c
@@ -760,7 +760,7 @@ int __pascal far get_tile_div_mod(int xpos) {
 			xh = bogus[COUNT(bogus) + xpos]; // simulating tile_div_tbl[xpos]
 			xl = tile_div_tbl[COUNT(tile_div_tbl) + xpos]; // simulating tile_mod_tbl[xpos]
 		} else {
-			printf("xpos = %d (< %ld) out of range for simulation of index overflow!\n", xpos, -COUNT(bogus));
+			printf("xpos = %d (< %d) out of range for simulation of index overflow!\n", xpos, -(int)COUNT(bogus));
 		}
 	}
 
@@ -775,7 +775,7 @@ int __pascal far get_tile_div_mod(int xpos) {
    xh = tile_mod_tbl[xpos-tblSize]; // simulating tile_div_tbl[xpos]
    xl = bogus[xpos-tblSize]; // simulating tile_mod_tbl[xpos]
   } else {
-   printf("xpos = %d (> %ld) out of range for simulation of index overflow!\n", xpos, COUNT(bogus)+tblSize);
+   printf("xpos = %d (> %d) out of range for simulation of index overflow!\n", xpos, (int)COUNT(bogus)+tblSize);
   }
  }
